Refactor Unit Test
